user-service/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
bcrypt
python-jose
python-multipart
pytest
httpx
PyJWT